Implement generic index functionality
