+++ kolibrie/examples/sparql_syntax/knowledge_graph/knowledge_graph.rs
@@ -11,6 +11,7 @@
 use shared::dictionary::Dictionary;
 use shared::terms::Term;
 use shared::rule::Rule;
+use shared::index_manager::TripleIndex;
 use datalog::reasoning::*;
 use datalog::parser_n3_logic::parse_n3_rule;
